Rework event structure
I've rewritten all Loom events to be more organized and easier to build onto. During the process, I removed quite a few unfinished or unimplemented events and kept only the events we can implement at this current moment. The result is a far simpler and more organized event list that should be a dream for developers to use.