Event encapsulation
This commit reworks the events API. Each event is now an interface in the API and the implementation takes care of all details. This allows us to mostly deprecate the event dispatcher and encapsulate events.

I also removed setCanceled and isCanceled from the parent Event class, since this makes it seem like EVERY event is cancelable, when in reality this is not the case.

A few other (minor) changes that made it into this commit whilst I was reworking events:
- Removed ItemEntityPlaceEvent since it was completely unimplemented
- Replaced SpongeEvent.Dry with a more generic BlockEvent.Dry
- Split EntityEvent.Damage into EntityEvent.Damage.Environmental and EntityEvent.Damage.Entity
- ChatEvent.Entity -> PlayerEvent.Chat since only players can "chat" (in the sense of sending a player chat message)
- CreeperEvent.Power -> CreeperEvent.Charge
- Fixed the implementation of StatusPingEvent